Remove old handlers and rename new handler folder
