+++ src/server/config/helpers.js
@@ -20,44 +20,5 @@ module.exports = {
   // Dummy function to test routes
   testy: function (req, res, next) {
     res.json("Route Succesful");
-  },
-  // Function to get users Git token from DB
-  getToken: function (gitID, cb) {
-    var tokQry = "SELECT git_token FROM users WHERE github_id='" +  gitID.toString() + "'";
-    pool.getConnection(function(err, connection){
-      if(err) {
-        throw err;
-      }
-      connection.query(tokQry, function (err, result) {
-        if (err) {
-          throw new Error(err);
-        } else {
-          cb(result[0].git_token);
-        }
-      });
-    });
-  },
-  // Function to grab Information from specified URL
-  gitURL: function (id, url, callback) {
-    var token;
-    console.log('Git ID: ', id);
-    module.exports.getToken(id,function(token){
-      token = token;
-      var options = {
-        url: url,
-        headers: {
-          'User-Agent': 'GitSpy',
-          authorization: 'token '+ token,
-          'content-type': 'application/json'
-        },
-      };
-      request.get(options, function(error, response, body) {
-        if (error){
-          throw new Error(error);
-        } else {
-          callback(response);
-        }
-      });
-    });
-  },
+  }
 };
