+++ src/server/request-handlers/repos.js
@@ -1,12 +1,18 @@
 "use strict";
-var helper = require('../config/helpers.js');
+var github = require('../queries/github.js');
 var request = require('request');
+var users = require('../queries/users.js');
 
 module.exports = {
   handlePost : function(req, res, next) {
     var gitID = req.cookies.githubId;
-    helper.gitURL(gitID, req.body.link, function(nextPage){
-      res.json(nextPage);
-    });
+    users.getOneAsync(gitID)
+      .then(function(usrObj){
+        var token = usrObj.token;
+        return github.queryAsync(req.body.link, usrObj.github_token);
+      })
+      .then(function(nextPage){
+        res.json(nextPage);
+      });
   }
 };
